+++ src/frontier_search.h
@@ -32,6 +32,6 @@ bool is_a_frontier_point(const pair<int,int> &p);
 vector<pair<int,int>> get_medians(vector<vector<pair<int,int>>> list_of_frontiers);
 bool has_open_neighbour(const vector<vector<int>> &marker_list, const pair<int,int> &p);
 vector<pair<int,int>> findNeighbours(const pair<int,int> &p);
-vector<pair<int,int>> wfd(tf::TransformListener &tf_listener);
+vector<pair<double,double>> wfd(tf::TransformListener &tf_listener);
 
 #endif
